// src/api/__generated__/models/User.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type User = {
    userId?: number;
    firstName?: string | null;
    lastName?: string | null;
    email?: string | null;
    password?: string | null;
    personalTrainerId?: number | null;
    accountType?: string | null;
}